Refactor the entire backend, details will be included in PR
