+++ backend/entites/protected_classes.py
@@ -1,0 +1,20 @@
+PROTECTED_CLASSES = {
+    "citizenship",
+    "sex",
+    "pregnancy",
+    "race",
+    "family status",
+    "place of origin",
+    "marital status",
+    "ethnic origin",
+    "sexual orientation",
+    "color",
+    "gender identity",
+    "ancestry",
+    "gender expression",
+    "disability",
+    "receipt of public assistance (in housing)",
+    "age",
+    "record of offenses (in employment)",
+    "creed"
+}
