+++ backend/tests/use_cases_tests.py
@@ -1,8 +1,6 @@
 import unittest
 import pandas as pd
 
-from useCases import helpers
-
 
 class TestHelperFunctions (unittest.TestCase):
     def setUp(self):
